# docs/swagger.yaml
components:
  schemas:
    fiber.Error:
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
...
    time.Duration:
      description: 재연결 시도 주기
      enum:
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
      type: integer
      x-enum-varnames:
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
    types.BadgerKeyRes:
      properties:
        keys:
          description: 매칭된 키 목록
          items:
            type: string
          type: array
          uniqueItems: false
        prefix:
          description: 조회 시 사용한 접두사
          type: string
      type: object
    types.CloudAuthRes:
      properties:
        url:
          type: string
      type: object
    types.CloudListRes:
      properties:
        items:
          items:
            type: string
          type: array
          uniqueItems: false
        path:
          type: string
      type: object
...
    types.LoginReq:
      properties:
        password:
          type: string
        username:
          type: string
      type: object
...
    types.SSEData:
      description: 데이터 페이로드
      properties:
        message:
          description: 메시지 내용
          type: string
        meta:
          $ref: '#/components/schemas/types.SSEMeta'
        status:
          description: 성공 여부
          type: boolean
        timestamp:
          description: 이벤트 발생 시간
          type: string
      type: object
    types.SSEEvent:
      description: 이벤트 유형
      enum:
      - heartbeat
      - subscribe
      - unsubscribe
      - publish
      - error
      type: string
      x-enum-varnames:
      - SSEEventHeartbeat
      - SSEEventSubscribe
      - SSEEventUnsubscribe
      - SSEEventPublish
      - SSEEventError
    types.SSEMessage:
      properties:
        data:
          $ref: '#/components/schemas/types.SSEData'
        event:
          $ref: '#/components/schemas/types.SSEEvent'
        id:
          description: 메시지 고유 ID
          type: string
        retry:
          $ref: '#/components/schemas/time.Duration'
      type: object
    types.SSEMeta:
      description: 메타데이터
      properties:
        action:
          description: 수행된 액션
          type: string
        id:
          description: 관련 리소스 ID
          type: string
        path:
          description: 관련 리소스 경로
          type: string
      type: object
    types.TokenRes:
      properties:
        expiresAt:
          type: string
        token:
          type: string
        username:
          type: string
      type: object
...
info:
  contact:
    email: miniyu97@gmail.com
    name: meteormin
  description: govfs is a virtual file system server
  title: govfs
  version: 1.0.0
openapi: 3.1.0
paths:
  /auth/login:
    post:
      description: login
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/types.LoginReq'
                description: login request
                summary: request
        description: login request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.TokenRes'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: logout
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Internal Server Error
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: is logged in
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.TokenRes'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Internal Server Error
      summary: IsLoggedIn
      tags:
      - auth
  /cloud/delete:
    delete:
      description: Deletes a file from the cloud storage.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: Delete file
      tags:
      - cloud
  /cloud/download:
    post:
      description: Downloads a file from the cloud storage.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/octet-stream:
              schema: {}
          description: OK
        "400":
          content:
            application/octet-stream:
              schema:
                type: string
          description: Bad Request
      summary: Download file
      tags:
      - cloud
  /cloud/googledrive/auth-code-url:
    get:
      description: Returns the authentication URL for Google Drive.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.CloudAuthRes'
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: Google Drive Auth Code URL
      tags:
      - cloud
  /cloud/googledrive/callback:
    get:
      description: Handles the callback from the Google Drive authentication.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: Google Drive Callback
      tags:
      - cloud
  /cloud/is-authorized:
    get:
      description: Checks if the cloud storage is authorized.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: Is Authorized
      tags:
      - cloud
  /cloud/list:
    get:
      description: Lists the files in the cloud storage.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.CloudListRes'
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: List files
      tags:
      - cloud
  /cloud/upload:
    post:
      description: Uploads a file to the cloud storage.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad Request
      summary: Upload file
      tags:
      - cloud
  /sse/:id/publish:
    post:
      description: Publish a Server-Sent Event message
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/types.SSEMeta'
                description: meta data
                summary: metaData
        description: meta data
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Bad Request
      summary: Publish SSE Message
      tags:
      - SSE
  /sse/subscribe:
    get:
      description: Subscribe to Server-Sent Events
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/types.SSEMessage'
          description: OK
        "500":
          content:
            text/event-stream:
              schema:
                type: object
          description: Internal Server Error
      summary: SSE Subscribe
      tags:
      - SSE
...
  /vfs/badger/keys:
    get:
      description: all keys
      parameters:
      - description: prefix
        in: query
        name: prefix
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.BadgerKeyRes'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fiber.Error'
          description: Internal Server Error
      summary: AllKeys
      tags:
      - vfs
...
servers:
- description: Localhost
  url: http://localhost:3000